requirements.txt
streamlit
gpxpy
folium
selenium
webdriver-manager
Pillow
imageio
